feat: Allow specifying warp scene name in config
The DebugWarp feature now accepts a scene name directly from the configuration file, removing the need to hardcode it in DebugWarp.cs. This allows users to easily toggle and choose the warp destination via the in-game configuration menu.

The configuration entry has been changed from a boolean to a string to accommodate the scene name. The README has also been updated to reflect this change.

(This commit message was automatically generated by an LLM.)